Refactor packages.
