+++ src/main/java/io/jking/tickster/cache/CachedObject.java
@@ -1,4 +1,4 @@
-package io.jking.tickster.objects.cache;
+package io.jking.tickster.cache;
 
 
 import io.jking.tickster.database.Database;
