+++ src/main/java/io/jking/tickster/cache/impl/GuildCache.java
@@ -1,8 +1,7 @@
-package io.jking.tickster.objects.cache.impl;
+package io.jking.tickster.cache.impl;
 
+import io.jking.tickster.cache.CachedObject;
 import io.jking.tickster.database.Database;
-import io.jking.tickster.objects.cache.CachedObject;
-import io.jking.untitled.jooq.tables.GuildData;
 import io.jking.untitled.jooq.tables.records.GuildDataRecord;
 import net.jodah.expiringmap.ExpirationPolicy;
 import org.jooq.Field;
@@ -10,7 +9,6 @@ import org.jooq.UpdateResultStep;
 
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
-import java.util.stream.Collectors;
 
 import static io.jking.untitled.jooq.tables.GuildData.GUILD_DATA;
