+++ src/main/java/io/jking/tickster/command/Category.java
@@ -1,4 +1,4 @@
-package io.jking.tickster.objects.command;
+package io.jking.tickster.command;
 
 public enum Category {
