+++ src/main/java/io/jking/tickster/command/impl/utility/PingCommand.java
@@ -1,9 +1,9 @@
-package io.jking.tickster.commands.utility;
+package io.jking.tickster.command.impl.utility;
 
-import io.jking.tickster.objects.command.Category;
-import io.jking.tickster.objects.command.Command;
-import io.jking.tickster.objects.command.CommandContext;
-import io.jking.tickster.objects.command.CommandError;
+import io.jking.tickster.command.Category;
+import io.jking.tickster.command.Command;
+import io.jking.tickster.command.CommandContext;
+import io.jking.tickster.command.CommandError;
 
 public class PingCommand extends Command {
     public PingCommand() {
